DBTrans: Implement AutoCloseable and create the connection yourself
Instead of the caller passing a Connection object into the constructor,
the caller must now pass a SQLiteDataSource so that we can create the
connection ourselves.

We close out the connection in our close method, following the rule that
whichever resource we create, we are responsible to destroy.

For now, the App method is not following that rule, because it's
creating a new DBTrans without calling its close() method, but we will
clean that up shortly.